fix: repairs multipacket requests
several incorrect semantics were discovered in the parse_body fun which
handles the case when the current length is incomplete.

1) the ++ operator was called to try to stitch together binaries.  This
was changed to the cons operator to make it an iolist.

2) the length() kernel function was called to try to measure the
length of the binary, this was changed to size()

3) the final piece putting together the htstub_request was changed to
convert the streamed body from iolist into a proper binary.  (this
might not actually be performant but whatever).

As a bonus, an extra guard was put in preventing idiots like me from
doing a stupid thing with htstub:rest